requirements.txt
# ── IMPORTANT: Python 3.12 required ──────────────────────────────────────────
# Python 3.14 is not yet supported by pandas, lxml (C extensions), or several
# dagster integration packages. Use Python 3.12:
#   winget install Python.Python.3.12
#   py -3.12 -m venv .venv && .venv\Scripts\activate
# ─────────────────────────────────────────────────────────────────────────────

# ── Core scraping ─────────────────────────────────────────────────────────────
Scrapy==2.11.*
scrapy-rotating-proxies==0.6.*
itemadapter==0.9.*
beautifulsoup4==4.12.*
lxml>=5.2,<6
requests==2.31.*
brotli>=1.1               # Required for Scrapy to decompress br (Brotli) responses

# ── Snowflake ─────────────────────────────────────────────────────────────────
snowflake-connector-python[pandas]>=3.7,<4
snowflake-sqlalchemy==1.5.*

# ── Orchestration (Dagster) ───────────────────────────────────────────────────
# dagster core: 1.X.Y  →  integration packages: 0.(X+16).Y
# dagster 1.12.*  →  dagster-dbt / dagster-snowflake / dagster-postgres 0.28.*
dagster==1.12.*
dagster-webserver==1.12.*
dagster-dbt==0.28.*
dagster-snowflake==0.28.*
dagster-postgres==0.28.*

# ── dbt ───────────────────────────────────────────────────────────────────────
dbt-core==1.8.*
dbt-snowflake==1.8.*

# ── Utilities ─────────────────────────────────────────────────────────────────
python-dotenv==1.0.*
pydantic==2.6.*
tenacity==8.2.*
Flask==3.0.*
stripe==11.6.*
reportlab==4.2.*
openpyxl==3.1.*

# ── Data (analysis notebooks — requires Python 3.12) ──────────────────────────
pandas==2.2.*
pyarrow>=15,<20
matplotlib>=3.8,<4

# ── Data quality ──────────────────────────────────────────────────────────────
great-expectations==0.18.*

# ── Testing ───────────────────────────────────────────────────────────────────
pytest==8.1.*
pytest-mock==3.12.*
responses==0.25.*
